+++ core/src/commonMain/kotlin/at/asitplus/awesn1/Asn1Exception.kt
@@ -22,9 +22,10 @@ class Asn1OidException(message: String, val oid: ObjectIdentifier) : Asn1Excepti
 /**
  * Runs [block] inside [catchingUnwrapped] and encapsulates any thrown exception in an [Asn1Exception] unless it already is one
  */
+@PublishedApi
 @Throws(Asn1Exception::class)
-inline fun <reified R> runRethrowing(block: () -> R) =
-    catchingUnwrapped(block).wrapAs(::Asn1Exception).getOrThrow()
+internal inline fun <R> runRethrowing(block: () -> R) =
+    runWrappingAs(a=::Asn1Exception, block)
 
 /**
  * Decodes this ASN.1 structure using the provided [decoder] lambda, rethrowing any caught exception
@@ -32,7 +33,7 @@ inline fun <reified R> runRethrowing(block: () -> R) =
  * This is a wrapper around [Asn1Structure.decodeAs] that ensures exceptions thrown during decoding are
  * consistently rethrown as [Asn1Exception], using the [runRethrowing] utility.
  */
-inline fun <reified R> Asn1Structure.decodeRethrowing(
+inline fun <R> Asn1Structure.decodeRethrowing(
     requireFullConsumption: Boolean = true,
     decoder: Asn1Structure.Iterator.() -> R
 ) = runRethrowing { this@decodeRethrowing.decodeAs(requireFullConsumption, decoder) }