+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/pki/Pkcs10CertificationRequest.kt
@@ -6,6 +6,8 @@ package at.asitplus.awesn1.crypto.pki
 import at.asitplus.awesn1.Asn1BitString
 import at.asitplus.awesn1.Asn1Element
 import at.asitplus.awesn1.Asn1Exception
+import at.asitplus.awesn1.Asn1PemDecodable
+import at.asitplus.awesn1.Asn1PemEncodable
 import at.asitplus.awesn1.Asn1Sequence
 import at.asitplus.awesn1.decodeRethrowing
 import at.asitplus.awesn1.encoding.Asn1
@@ -19,9 +21,9 @@ open class Pkcs10CertificationRequest(
     val certificationRequestInfo: Pkcs10CertificationRequestInfo,
     val signatureAlgorithm: SignatureAlgorithmIdentifier,
     val signatureValue: SignatureValue,
-) : at.asitplus.awesn1.Asn1PemEncodable<Asn1Sequence> {
+) : Asn1PemEncodable<Asn1Sequence> {
 
-    override val pemLabel: String = "CERTIFICATE REQUEST"
+    override val pemLabel get() = PEM_LABEL
 
     override fun encodeToTlv() = Asn1.Sequence {
         +certificationRequestInfo
@@ -44,8 +46,14 @@ open class Pkcs10CertificationRequest(
         return result
     }
 
-    companion object : Asn1Serializable<Asn1Sequence, Pkcs10CertificationRequest> {
+    companion object :
+        Asn1Serializable<Asn1Sequence, Pkcs10CertificationRequest>,
+        Asn1PemDecodable<Asn1Sequence, Pkcs10CertificationRequest>
+    {
         override val leadingTags = setOf(Asn1Element.Tag.SEQUENCE)
+
+        const val PEM_LABEL = "CERTIFICATE REQUEST"
+        override val pemLabel get() = PEM_LABEL
 
         @Throws(Asn1Exception::class)
         override fun doDecode(src: Asn1Sequence): Pkcs10CertificationRequest = src.decodeRethrowing {